+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Random first sleep in background thread
 * Refactor of usage of Thread.local
 * Net::HTTP instrumentation
+* Capistrano 3 support
 
 # 0.8.10
 * Bugfix: Don't pause agent when it's not active